[REFACTOR] 인기있는 소설을 Application 계층으로 이동
