+++ src/main/java/org/websoso/WSSServer/application/SearchNovelApplication.java
@@ -5,48 +5,46 @@ import static org.websoso.WSSServer.exception.error.CustomGenreError.GENRE_NOT_F
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collections;
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Random;
+import java.util.Set;
 import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.websoso.WSSServer.domain.Avatar;
 import org.websoso.WSSServer.domain.Genre;
 import org.websoso.WSSServer.domain.GenrePreference;
 import org.websoso.WSSServer.domain.User;
-import org.websoso.WSSServer.domain.common.AttractivePointName;
 import org.websoso.WSSServer.domain.common.GenreName;
-import org.websoso.WSSServer.dto.keyword.KeywordCountGetResponse;
 import org.websoso.WSSServer.dto.novel.FilteredNovelsGetResponse;
 import org.websoso.WSSServer.dto.novel.NovelGetResponseBasic;
 import org.websoso.WSSServer.dto.novel.NovelGetResponseInfoTab;
 import org.websoso.WSSServer.dto.novel.NovelGetResponsePreview;
 import org.websoso.WSSServer.dto.novel.SearchedNovelsGetResponse;
+import org.websoso.WSSServer.dto.popularNovel.PopularNovelsGetResponse;
 import org.websoso.WSSServer.dto.userNovel.TasteNovelGetResponse;
 import org.websoso.WSSServer.dto.userNovel.TasteNovelsGetResponse;
 import org.websoso.WSSServer.exception.exception.CustomGenreException;
+import org.websoso.WSSServer.feed.domain.Feed;
 import org.websoso.WSSServer.feed.repository.FeedRepository;
 import org.websoso.WSSServer.library.domain.Keyword;
 import org.websoso.WSSServer.library.domain.UserNovel;
-import org.websoso.WSSServer.library.domain.UserNovelKeyword;
 import org.websoso.WSSServer.library.service.LibraryService;
 import org.websoso.WSSServer.novel.domain.Novel;
 import org.websoso.WSSServer.novel.domain.NovelGenre;
 import org.websoso.WSSServer.novel.service.GenreServiceImpl;
 import org.websoso.WSSServer.novel.service.KeywordServiceImpl;
 import org.websoso.WSSServer.novel.service.NovelServiceImpl;
+import org.websoso.WSSServer.repository.AvatarRepository;
 import org.websoso.WSSServer.repository.GenrePreferenceRepository;
 
 @Service
 @RequiredArgsConstructor
 public class SearchNovelApplication {
-
-    private static final int ATTRACTIVE_POINT_SIZE = 3;
-    private static final int KEYWORD_SIZE = 5;
 
     private final NovelServiceImpl novelService;
     private final GenreServiceImpl genreService;
@@ -56,6 +54,7 @@ public class SearchNovelApplication {
     // TODO: 삭제될 레포지토리 의존성
     private final FeedRepository feedRepository;
     private final GenrePreferenceRepository genrePreferenceRepository;
+    private final AvatarRepository avatarRepository;
 
     /**
      * 검색어(소셜명, 작가명)에 해당하는 소설 찾기
@@ -129,8 +128,8 @@ public class SearchNovelApplication {
         return NovelGetResponseInfoTab.of(
                 novel,
                 novelService.getPlatforms(novel),
-                getAttractivePoints(novel),
-                getKeywords2(novel),
+                libraryService.getAttractivePoints(novel),
+                libraryService.getKeywordNameAndCount(novel),
                 libraryService.getWatchingCount(novel),
                 libraryService.getWatchedCount(novel),
                 libraryService.getQuitCount(novel)
@@ -154,6 +153,20 @@ public class SearchNovelApplication {
         return TasteNovelsGetResponse.of(tasteNovelGetResponses);
     }
 
+    @Transactional(readOnly = true)
+    public PopularNovelsGetResponse getTodayPopularNovels() {
+        List<Long> novelIdsFromPopularNovel = novelService.getNovelIdsFromPopularNovel();
+        List<Long> selectedNovelIdsFromPopularNovel = getSelectedNovelIdsFromPopularNovel(novelIdsFromPopularNovel);
+        List<Novel> popularNovels = novelService.getSelectedPopularNovels(selectedNovelIdsFromPopularNovel);
+        List<Feed> popularFeedsFromPopularNovels = feedRepository.findPopularFeedsByNovelIds(selectedNovelIdsFromPopularNovel);
+
+        Map<Long, Feed> feedMap = createFeedMap(popularFeedsFromPopularNovels);
+        Map<Byte, Avatar> avatarMap = createAvatarMap(feedMap);
+
+        return PopularNovelsGetResponse.create(popularNovels, feedMap, avatarMap);
+    }
+
+    // TODO: DTO로 이전할 명분이 충분한 메서드
     private NovelGetResponsePreview convertToDTO(Novel novel) {
         // TODO: UserNovel 리스트 개수를 세는것이 아닌, 개수를 세는 쿼리가 필요
         List<UserNovel> userNovels = novel.getUserNovels();
@@ -213,23 +226,6 @@ public class SearchNovelApplication {
         return keywords;
     }
 
-    private List<KeywordCountGetResponse> getKeywords2(Novel novel) {
-        List<UserNovelKeyword> userNovelKeywords = libraryService.getKeywords(novel);
-
-        if (userNovelKeywords.isEmpty()) {
-            return Collections.emptyList();
-        }
-
-        Map<Keyword, Long> keywordFrequencyMap = userNovelKeywords.stream()
-                .collect(Collectors.groupingBy(UserNovelKeyword::getKeyword, Collectors.counting()));
-
-        return keywordFrequencyMap.entrySet().stream()
-                .sorted(Map.Entry.<Keyword, Long>comparingByValue().reversed())
-                .limit(KEYWORD_SIZE)
-                .map(entry -> KeywordCountGetResponse.of(entry.getKey(), entry.getValue().intValue()))
-                .toList();
-    }
-
     private String getNovelGenreNames(List<NovelGenre> novelGenres) {
         return novelGenres.stream()
                 .map(novelGenre -> getKoreanGenreName(novelGenre.getGenre().getGenreName()))
@@ -250,64 +246,27 @@ public class SearchNovelApplication {
         return novelGenres.get(random.nextInt(novelGenres.size())).getGenre().getGenreImage();
     }
 
-
-    private List<String> getAttractivePoints(Novel novel) {
-        Map<String, Integer> attractivePointMap = makeAttractivePointMapExcludingZero(novel);
-
-        if (attractivePointMap.isEmpty()) {
-            return Collections.emptyList();
-        }
-
-        return getTOP3AttractivePoints(attractivePointMap);
+    private List<Long> getSelectedNovelIdsFromPopularNovel(List<Long> popularNovelIds) {
+        Collections.shuffle(popularNovelIds);
+        return popularNovelIds.size() > 10
+                ? popularNovelIds.subList(0, 10)
+                : popularNovelIds;
     }
 
-    private Map<String, Integer> makeAttractivePointMapExcludingZero(Novel novel) {
-        Map<String, Integer> attractivePointMap = new HashMap<>();
-
-        for (AttractivePointName point : AttractivePointName.values()) {
-            attractivePointMap.put(point.getLabel(), libraryService.getAttractivePointCount(novel, point));
-        }
-
-        attractivePointMap.entrySet().removeIf(entry -> entry.getValue() == 0);
-
-        return attractivePointMap;
+    private Map<Long, Feed> createFeedMap(List<Feed> popularFeedsFromPopularNovels) {
+        return popularFeedsFromPopularNovels.stream()
+                .collect(Collectors.toMap(Feed::getNovelId, feed -> feed));
     }
 
-    private List<String> getTOP3AttractivePoints(Map<String, Integer> attractivePointMap) {
-        Map<Integer, List<String>> groupedByValue = groupAttractivePointByValue(attractivePointMap);
+    private Map<Byte, Avatar> createAvatarMap(Map<Long, Feed> feedMap) {
+        Set<Byte> avatarIds = feedMap.values()
+                .stream()
+                .map(feed -> feed.getUser().getAvatarId())
+                .collect(Collectors.toSet());
 
-        List<String> result = new ArrayList<>();
-        List<Integer> sortedKeys = new ArrayList<>(groupedByValue.keySet());
-        sortedKeys.sort(Collections.reverseOrder());
-
-        Random random = new Random();
-
-        for (Integer key : sortedKeys) {
-            List<String> items = groupedByValue.get(key);
-            if (result.size() + items.size() > ATTRACTIVE_POINT_SIZE) {
-                Collections.shuffle(items, random);
-                items = items.subList(0, ATTRACTIVE_POINT_SIZE - result.size());
-            }
-            result.addAll(items);
-            if (result.size() >= ATTRACTIVE_POINT_SIZE) {
-                break;
-            }
-        }
-
-        return result;
+        List<Avatar> avatars = avatarRepository.findAllById(avatarIds);
+        return avatars.stream()
+                .collect(Collectors.toMap(Avatar::getAvatarId, avatar -> avatar));
     }
-
-    private Map<Integer, List<String>> groupAttractivePointByValue(Map<String, Integer> attractivePointMap) {
-        Map<Integer, List<String>> groupedByValue = new HashMap<>();
-
-        for (Map.Entry<String, Integer> entry : attractivePointMap.entrySet()) {
-            groupedByValue
-                    .computeIfAbsent(entry.getValue(), k -> new ArrayList<>())
-                    .add(entry.getKey());
-        }
-
-        return groupedByValue;
-    }
-
 
 }
