+++ src/main/java/org/websoso/WSSServer/novel/service/NovelServiceImpl.java
@@ -2,6 +2,8 @@ package org.websoso.WSSServer.novel.service;
 
 import static org.websoso.WSSServer.exception.error.CustomNovelError.NOVEL_NOT_FOUND;
 
+import java.util.ArrayList;
+import java.util.Collections;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
@@ -9,6 +11,7 @@ import org.springframework.data.domain.PageRequest;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.websoso.WSSServer.domain.Genre;
+import org.websoso.WSSServer.novel.domain.PopularNovel;
 import org.websoso.WSSServer.dto.platform.PlatformGetResponse;
 import org.websoso.WSSServer.exception.exception.CustomNovelException;
 import org.websoso.WSSServer.library.domain.Keyword;
@@ -17,6 +20,7 @@ import org.websoso.WSSServer.novel.domain.NovelGenre;
 import org.websoso.WSSServer.novel.repository.NovelGenreRepository;
 import org.websoso.WSSServer.novel.repository.NovelPlatformRepository;
 import org.websoso.WSSServer.novel.repository.NovelRepository;
+import org.websoso.WSSServer.novel.repository.PopularNovelRepository;
 
 @Service
 @RequiredArgsConstructor
@@ -25,6 +29,7 @@ public class NovelServiceImpl {
     private final NovelRepository novelRepository;
     private final NovelGenreRepository novelGenreRepository;
     private final NovelPlatformRepository novelPlatformRepository;
+    private final PopularNovelRepository popularNovelRepository;
 
     @Transactional(readOnly = true)
     public Novel getNovelOrException(Long novelId) {
@@ -51,6 +56,17 @@ public class NovelServiceImpl {
         return novelPlatformRepository.findAllByNovel(novel).stream()
                 .map(PlatformGetResponse::of)
                 .toList();
+    }
+
+    public List<Long> getNovelIdsFromPopularNovel() {
+        return new ArrayList<>(popularNovelRepository.findAll()
+                .stream()
+                .map(PopularNovel::getNovelId)
+                .toList());
+    }
+
+    public List<Novel> getSelectedPopularNovels(List<Long> selectedPopularNovelIds) {
+        return novelRepository.findAllById(selectedPopularNovelIds);
     }
 
 }
